fix linting

+++ pkg/rtpaac/decoder.go
@@ -48,9 +48,9 @@ func (d *Decoder) Decode(pkt *rtp.Packet) ([][]byte, time.Duration, error) {
 	}
 
 	// AU-headers-length
-	headersLen := binary.BigEndian.Uint16(pkt.Payload)
+	headersLen := int(binary.BigEndian.Uint16(pkt.Payload))
 
-	auHeaderSize := uint16(d.SizeLength + d.IndexLength)
+	auHeaderSize := d.SizeLength + d.IndexLength
 	if auHeaderSize <= 0 {
 		d.fragmentedParts = d.fragmentedParts[:0]
 		d.fragmentedMode = false
@@ -62,7 +62,7 @@ func (d *Decoder) Decode(pkt *rtp.Packet) ([][]byte, time.Duration, error) {
 		d.fragmentedMode = false
 		return nil, 0, fmt.Errorf("invalid AU-headers-length (%d) with AU-header-size (%d)", headersLen, auHeaderSize)
 	}
-	headersLenBytes := (int(headersLen) + 7) / 8
+	headersLenBytes := (headersLen + 7) / 8
 	payload := pkt.Payload[2:]
 
 	if !d.fragmentedMode {
@@ -162,14 +162,14 @@ func (d *Decoder) Decode(pkt *rtp.Packet) ([][]byte, time.Duration, error) {
 
 func (d *Decoder) parseAuData(payload []byte,
 	headersLenBytes int,
-	headerCount uint16,
+	headerCount int,
 ) (dataLens []uint64, err error) {
 	if len(payload) < headersLenBytes {
 		return nil, fmt.Errorf("payload is too short")
 	}
 
 	br := bitio.NewReader(bytes.NewBuffer(payload[:headersLenBytes]))
-	for i := 0; i < int(headerCount); i++ {
+	for i := 0; i < headerCount; i++ {
 		dataLen, err := br.ReadBits(uint8(d.SizeLength))
 		if err != nil {
 			return nil, fmt.Errorf("payload is too short")
